requirements.txt
anthropic>=0.18.1
python-dotenv>=1.0.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
flask>=3.0.0 